+++ package.json
@@ -28,9 +28,10 @@
     "update dependencies": "npm run check && npm run update",
     "lint": "eslint src/**.ts",
     "jlint": "eslint homebridge-ui/public/**.mjs",
-    "watch": "npm run build && npm link && nodemon",
+    "watch": "npm run build && npm run plugin-ui && npm link && nodemon",
+    "plugin-ui": "rsync ./src/homebridge-ui/public/index.html ./dist/homebridge-ui/public/",
     "build": "rimraf ./dist && tsc",
-    "prepublishOnly": "npm run lint && npm run build",
+    "prepublishOnly": "npm run lint && npm run build && npm run plugin-ui ",
     "postpublish": "npm run clean",
     "clean": "rimraf ./dist",
     "test": "eslint src/**.ts"
@@ -51,16 +52,16 @@
     "No-IP"
   ],
   "dependencies": {
-    "@homebridge/plugin-ui-utils": "^1.0.0",
-    "undici": "^6.2.1",
+    "@homebridge/plugin-ui-utils": "^1.0.1",
+    "undici": "^6.4.0",
     "rxjs": "^7.8.1"
   },
   "devDependencies": {
-    "@types/node": "^20.10.5",
-    "@typescript-eslint/eslint-plugin": "^6.15.0",
-    "@typescript-eslint/parser": "^6.15.0",
+    "@types/node": "^20.11.5",
+    "@typescript-eslint/eslint-plugin": "^6.19.0",
+    "@typescript-eslint/parser": "^6.19.0",
     "eslint": "^8.56.0",
-    "nodemon": "^3.0.2",
+    "nodemon": "^3.0.3",
     "homebridge": "^1.7.0",
     "npm-check-updates": "^16.14.12",
     "rimraf": "^5.0.5",
